pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-ulid>=3.0.0",
    "redis>=6.1.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ignr>=2.2",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.11",
    "taskipy>=1.14.1",
]


[tool.ruff]
line-length = 79
exclude = ["app/database/migrations"]

[tool.ruff.lint]
preview = true
ignore = ['PLR6301']
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
dev = 'fastapi dev projeto_aplicado/app.py'
test = 'pytest -s -x --cov=projeto_aplicado'
run = 'fastapi run projeto_aplicado/app.py'
export = 'uv pip compile pyproject.toml --output-file requirements.txt'
